Require Rails 4.2.0
