+++ Gemfile
@@ -2,6 +2,10 @@ source 'https://rubygems.org'
 
 gemspec
 
-gem 'railties', '~> 4.1.0'
-gem 'activemodel', '~> 4.1.0'
+# gem 'railties', '~> 4.2.0-alpha'
+# gem 'activemodel', '~> 4.2.0-alpha'
+gem 'rails', github: 'rails/rails'
+gem 'i18n', github: 'svenfuchs/i18n'
+gem 'rack', github: 'rack/rack'
+gem 'arel', github: 'rails/arel'
 gem 'mocha'
