--- gemfiles/Gemfile-rails.3.2.x
@@ -1,7 +1,0 @@
-source 'https://rubygems.org'
-
-gemspec path: '..'
-
-# Fixed 3.2.3 because 3.2.2 is broken.
-gem 'railties', '~> 3.2.3'
-gem 'mocha'
