--- gemfiles/Gemfile-rails.4.0.x
@@ -1,7 +1,0 @@
-source 'https://rubygems.org'
-
-gemspec path: '..'
-
-gem 'railties', '~> 4.0.0'
-gem 'activemodel', '~> 4.0.0'
-gem 'mocha'
