+++ lib/responders/version.rb
@@ -1,3 +1,3 @@
 module Responders
-  VERSION = "1.1.0".freeze
+  VERSION = "2.0.0".freeze
 end
